requirements.txt
aiohttp==3.8.5
colorama==0.4.6
pydantic==2.4.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
